+++ mbgdml/active_learning/drive_md.py
@@ -24,9 +24,14 @@
 
 import os
 import shutil
+import sys
 import uuid
 from ase import Atoms
+from ase import units as ase_units
+from ase.md.npt import NPT
+from ase.md.velocitydistribution import MaxwellBoltzmannDistribution
 from ase.optimize import BFGS
+from ase.io.trajectory import Trajectory
 import numpy as np
 import ray
 from ..interfaces.ase import mbeCalculator
@@ -49,11 +54,19 @@ class MDDriver:
         mol_comp_ids,
         mbe_pred,
         label=None,
+        periodic_cell=None,
         ase_e_conv=1.0,
         ase_f_conv=1.0,
         opt_f_max=0.4,
-        opt_steps=50,
-        periodic_cell=None,
+        opt_max_steps=50,
+        md_n_steps=1000,
+        md_t_step=1.0 * ase_units.fs,
+        md_init_temp=100.0,
+        md_temp=300.0,
+        nose_hoover_ttime=50.0 * ase_units.fs,
+        md_external_stress=101325 * ase_units.Pascal,
+        md_traj_interval=1,
+        md_stability_interval=5,
         keep_files=False,
     ):
         r"""
@@ -65,25 +78,56 @@ class MDDriver:
             Input file for packmol structure generation.
         mol_n_atoms : :obj:`tuple`
             Number of atoms for each "structure" line in the order they appear in
-            ``packmol_input``. Used to create ``entity_ids``
+            ``packmol_input``. Used to create ``entity_ids``.
+        mol_comp_ids : :obj:`tuple`
+            Component IDs of each species in the order that they appear in
+            ``packmol_input``. Used to create ``comp_ids``.
         mbe_pred : :obj:`mbgdml.mbe.mbePredict`
             Predictor function for driving MD simulations.
         label : :obj:`str`, default: ``None``
             Unique label for this MD simulation. Otherwise, a random label will be
             generated.
-        keep_files : :obj:`bool`, default: ``False``
+        periodic_cell : :obj:`mbgdml.periodic.Cell`, default: :obj:`None`
+            Use periodic boundary conditions defined by this object. If this
+            is not :obj:`None` only :meth:`~mbgdml.mbe.mbePredict.predict` can be used.
+        ase_e_conv : :obj:`float`, default: ``1.0``
+            Model energy conversion factor to eV (required by ASE).
+        ase_f_conv : :obj:`float`, default: ``1.0``
+            Model forces conversion factor to eV/A (required by ASE).
+        opt_f_max : :obj:`float`, default: ``0.4``
+            Max force in eV/A to be considered converged.
+        opt_max_steps : :obj:`int`, default: ``50``
+            Maximum number of geometry optimization steps.
+        md_n_steps : :obj:`int`, default: ``1000``
+            Number of MD steps to take after optimization.
+        md_t_step : :obj:`float`, default: ``1.0*ase_units.fs``
+            MD time step in ASE units.
+        md_init_temp : :obj:`float`, default: ``100``
+            Temperature used to initialize velocities for MD simulations using the
+            Maxwell-Boltzmann distribution.
+        md_temp : :obj:`float`, default: ``300``
+            Temperature set point for thermostat.
+        nose_hoover_ttime : :obj:`float`, default: ``50.0*ase_units.fs``
+            Characteristic timescale of the thermostat in ASE internal units.
+        md_external_stress : :obj:`float`, default: ``101325*ase_units.Pascal``
+            External stress to the system in ASE internal units. The value is not
+            actually used.
+        md_traj_interval : :obj:`int`, default: ``1``
+            Steps between logging in the trajectory.
+        keep_files : :obj:`bool`, default: ``True``
             Keep all working files of this Actor.
         """
         self.log = GDMLLogger(__name__)
         self.log.info("Initializing MDDriver")
+
         self.work_dir = work_dir
         if not os.path.exists(work_dir):
             raise RuntimeError(f"The working directory, {work_dir}, does not exist")
 
         self.packmol_input = packmol_input
-
         self.mol_n_atoms = mol_n_atoms
         self.mol_comp_ids = mol_comp_ids
+
         self.mbe_pred = mbe_pred
 
         if label is None:
@@ -93,14 +137,24 @@ class MDDriver:
         self.temp_dir = os.path.join(work_dir, label)
         os.mkdir(self.temp_dir)
 
-        self.keep_files = keep_files
-
+        self.periodic_cell = periodic_cell
         self.ase_e_conv = ase_e_conv
         self.ase_f_conv = ase_f_conv
         self.opt_f_max = opt_f_max
-        self.opt_steps = opt_steps
-        self.periodic_cell = periodic_cell
-        self.opt_traj_name = "geom-opt.traj"
+        self.opt_max_steps = opt_max_steps
+        self.md_n_steps = md_n_steps
+        self.md_t_step = md_t_step
+        self.md_init_temp = md_init_temp
+        self.md_temp = md_temp
+        self.nose_hoover_ttime = nose_hoover_ttime
+        self.md_external_stress = md_external_stress
+        self.md_traj_interval = md_traj_interval
+        self.md_stability_interval = md_stability_interval
+
+        self.opt_traj_name = "opt.traj"
+        self.md_traj_name = "nvt.traj"
+
+        self.keep_files = keep_files
 
         self.log.info("Done initializing MDDriver")
 
@@ -120,17 +174,7 @@ class MDDriver:
 
     # pylint: disable-next=invalid-name
     def create_starting_R(self):
-        r"""Create the starting structure for a MD simulation using packmol.
-
-        Parameters
-        ----------
-        packmol_input : :obj:`str`
-            Input file for packmol structure generation.
-        mol_n_atoms : :obj:`tuple`
-            Number of atoms for each "structure" line in the order they appear in
-            ``packmol_input``. Used to create ``entity_ids``
-
-        """
+        r"""Create the starting structure for a MD simulation using packmol."""
         self.log.info("Generating structure with packmol")
 
         # Determine entity_ids and comp_ids for predictions.
@@ -155,15 +199,7 @@ class MDDriver:
         self.Z_packmol, self.R_packmol = run_packmol(self.temp_dir, self.packmol_input)
 
     def setup_ase(self):
-        """
-
-        Parameters
-        ----------
-        e_conv : :obj:`float`, default: ``1.0``
-            Model energy conversion factor to eV (required by ASE).
-        f_conv : :obj:`float`, default: ``1.0``
-            Model forces conversion factor to eV/A (required by ASE).
-        """
+        """Prepares ASE atoms with calculator."""
         self.log.info("Setting up ASE calculator")
         pbc = False
         cell_v = None
@@ -181,37 +217,98 @@ class MDDriver:
         )
         self.ase_atoms.calc = mbe_calc
 
-    def optimize(self, traj_path="geom-opt.traj"):
+    def optimize(self):
         r"""Perform geometry optimization on ASE Atoms object."""
         self.log.info("Performing geometry optimization")
         traj_path = os.path.join(self.temp_dir, self.opt_traj_name)
         dyn = BFGS(atoms=self.ase_atoms, trajectory=traj_path)
-        dyn.run(fmax=self.opt_f_max, steps=self.opt_steps)
+        dyn.run(fmax=self.opt_f_max, steps=self.opt_max_steps)
 
-    def check_md_stability(self, R, E):
-        r"""Determines if the MD simulation is unstable and needs to be terminated.
-
-        Parameters
-        ----------
-        R : :obj:`numpy.ndarray`
-            Cartesian coordinates of the last :math:`N` steps.
-        E : :obj:`numpy.ndarray`
-            Energies of the last :math:`N` steps. Can be potential, kinetic, or total
-            energies.
+    def check_md_stability(self):
+        r"""Determines if the MD simulation is stable. If not, it will be terminated.
+        
+        Returns
+        -------
+        :obj:`bool`
+            If the MD simulation is stable.
         """
+        self.log.info("Checking MD simulations")
+        traj_path = os.path.join(self.temp_dir, self.md_traj_name)
+        md_traj_read = Trajectory(traj_path, mode="r")
+        energies = []
+        for atoms in md_traj_read[-self.md_stability_interval :]:
+            energies.append(atoms.get_potential_energy())
+        energies = np.array(energies)
+        energies -= energies[0]
+        energy_change_abs_mean = np.abs(np.mean(energies[1:]))
+        self.log.debug("Average absolute energy change: %d", energy_change_abs_mean)
+        threshold = 3.0
+        if energy_change_abs_mean > threshold:
+            self.log.debug("This is above the threshold of %d", threshold)
+            self.log.debug("MD simulation is considered unstable")
+            return False
+        self.log.debug("MD simulation is considered stable")
+        return True
 
-    def step(self):
-        """Perform a single MD step."""
+    def nvt(self):
+        """Perform MD simulation."""
+        self.log.info("Setting up MD")
+        ase_atoms = self.ase_atoms
+        self.log.debug("Initializing velocities using Maxwell-Boltzmann distribution")
+        MaxwellBoltzmannDistribution(ase_atoms, temperature_K=self.md_init_temp)
+        self.log.debug("Setting up NPT object")
+        self.md = NPT(  # pylint: disable=invalid-name
+            atoms=ase_atoms,
+            timestep=self.md_t_step,
+            externalstress=self.md_external_stress,
+            ttime=self.nose_hoover_ttime,
+            pfactor=None,
+            temperature_K=self.md_temp,
+        )
+        self.log.debug("Setting up loggers")
+        md_traj_path = os.path.join(self.temp_dir, self.md_traj_name)
+        md_traj = Trajectory(md_traj_path, mode="w", atoms=ase_atoms)
+
+        self.md_step = 0
+
+        def print_ase_step(
+            a=ase_atoms,
+        ):  # store a reference to atoms in the definition.
+            """Function to print status of MD simulation."""
+            epot = a.get_potential_energy()
+            ekin = a.get_kinetic_energy()
+            ekin_per_atom = ekin / len(a)
+            temp_step = ekin_per_atom / (1.5 * ase_units.kB)
+
+            print(
+                f"Step {self.md_step : >5}/{self.md_n_steps : <5}     "
+                f"E_pot: {epot : >15.8f} eV     "
+                f"T: {temp_step : >7.2f} K"
+            )
+            self.md_step += 1
+            if self.md_step > self.md_stability_interval:
+                if not self.check_md_stability():
+                    # Raises caught system exit to terminate MD simulation.
+                    sys.exit("MD simulation is unstable")
+
+        self.md.attach(print_ase_step, interval=1)
+        self.md.attach(md_traj.write, interval=self.md_traj_interval)
+        self.log.info("Starting simulation")
+        try:
+            self.md.run(self.md_n_steps)
+        except SystemExit as e:
+            if "MD simulation is unstable" != str(e):
+                raise e
 
     def cleanup(self):
-        r"""Cleanup Actor files and such."""
+        r"""Cleanup temporary files and such."""
         if not self.keep_files:
             shutil.rmtree(self.temp_dir)
 
     def run(self):
-        """ """
+        r"""Prepare and run MD simulation."""
         self.create_starting_R()
         self.setup_ase()
         self.optimize()
+        self.nvt()
         self.cleanup()
-        return True